requirements.txt
# ChatOps
aiogram>=3.4,<4
discord.py>=2.3,<3
python-dotenv>=1.0

# Core agent dependencies
openai>=1.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.2
tweepy>=4.16.0
praw>=7.7.1
Mastodon.py>=1.8.0
atproto>=0.0.36
Pillow>=10.0.0
rich>=13.5.2
humanize>=4.8.0
